Add check for file existence in saveJSON method (The thing is on my computer, tests are passing, on GitHub - they are not)

+++ src/main/java/io/github/dayfit/JSON.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import java.io.IOException;
 import java.nio.file.Files;
+import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
 
 public class JSON {
@@ -24,6 +25,10 @@ public class JSON {
     }
 
     public static void saveJSON(String json, Path path) throws IOException {
+        if (!Files.exists(path))
+        {
+            throw new NoSuchFileException("No such file or directory");
+        }
         Files.write(path, json.getBytes());
     }
 }
